requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyMySQL==1.1.1
WTForms==3.1.2
Flask-WTF==1.2.1
python-dotenv==1.0.1
Werkzeug==3.0.3
alembic==1.13.2
requests==2.32.3
cryptography==43.0.1
gunicorn==21.2.0
